requirements.txt
pytorch-gpu==1.3.1
tensorflow-gpu==1.14.0
scikit-learn
scikit-image
torchvision
tqdm 
pyyaml
cloudpickle
ipython
opencv